requirements.txt
prefect>=2.14,<3
google-cloud-storage>=2.14
google-cloud-bigquery>=3.14
pandas>=2.0
pyarrow>=14.0
python-dotenv>=1.0
dbt-bigquery>=1.7,<2
streamlit>=1.30
requests>=2.31
